replaced gson with jackson

+++ banshie-api/src/main/java/org/whiskeysierra/banshie/evaluation/Value.java
@@ -1,8 +1,8 @@
 package org.whiskeysierra.banshie.evaluation;
 
+// TODO min, max, average?
 public interface Value {
 
-    // TODO consider better number type!
-    float getValue();
+    double getValue();
 
 }
